Fixed a bug which was introduced in the revision #759. Mapster32 would delete many sprites on saving maps.
The latest released snapshot is affected by this bug. This is a critical bug and a new snapshot should be released as soon as possible and users should be asked to not use the corrupted snapshot.

The bug was reported by DeeperThought and Gambini.

P.S. Actually it's the second time this bug has been introduced. See http://forums.3drealms.com/vb/showthread.php?t=29413

From-SVN: r772